refactor: move block production to its crate
