+++ crates/client/block_production/src/close_block.rs
